Adding basic support for malloc
Implementing _sbrk which is used for malloc.

The startup file has been extended to include initializing the .bss
section to zero. This is EXTREMELY important becuase malloc assumes
its global variables are intiliazed to 0. If this is not done, then
variables in that section will have undefined behaviour, most of which
result in a bus fault.

Changing heap name from _pvHeapStart to _heap.