Refactoring repo and adding clang-format and lint
All source files have been refactored to follow the CPPLINT.cfg file and
the .clang-format file.

macros.hpp was added to simply and standardize the use of (void) to
suppress warnings and adding symbols into sections. All macros made sj2
must be prefixed with SJ2_. Private _macros that should not be used
outside of libraries files should be prefixed with _SJ2_.