Adding backtrace support
The desired effect is to allow assert to print the call stack so the user
can determine what sequence of function calls caused the assert to occur.

Removing unnecessary keep-folder.txt files

Does not work with faults and interrupt, may add that functionality at a
later time.

Light sprucing up of code cruft.