Changing name of PinConfigure -> Pin
PinConfigure is a bulky and large name to write as well as to read and
say out loud. Its means seems a bit confusing. Simply calling it a Pin
makes more sense.

It also makes Gpio's inheritance of Pin make more sense.

Gpio "is a" pin VS Gpio "is a" PinConfigure

Where the later doesn't make a whole lot of sense.