Adding runtime task location check to Task class
The only two places a task can exist is in the .data section or the
.heap, or any memory that is presistant.

If made on the stack, this will most likely lead to a hard fault.

To prevent developers from making this mistake, during construction,
tasks do a self test of their own address and make sure they are within
the bounds of the .data, .bss, or .heap sections. If this fails, a fatal
assertion is triggered the the program will halt. Othewise, the program
will continue.