Move from config::kSystemFreq -> GetSystemFreq()
Also added a new SystemController method GetPeripheralFrequency() which
calculates the clock rate provided to AHB bus using the PCLKSEL register
and the previously set system clock frequency.

It is prefered that L1 drivers use this rather than using the
config::kSystemFrequency or GetSystemFrequency().
config::kSystemFrequency only reflects the target frequency
that the system wants to achieve and not what it actually is, and
GetSystemFrequency() does not tell you what clock the peripheral is
running at.