Optimized I2C Handler template
Reduced example/I2C from 34688 bytes -> 32192 bytes (2496 bytes)
by implementing the templated I2CHandler as a handoff to a static
I2CHandler that takes a port argument.

This is more efficent since the statemachine doesn't have to be created
for each i2c port, but simply a function that passes the port number for
that i2c port into the static handler.