Linking of SystemTimer no longer universal
SystemTimer, before the integration of the Timer peripheral, was used
for delay timing. Now the timer0 object exists, so there is no longer
any need to link in SystemTimer unless FreeRTOS is used.

Now SystemTimer is a local variable to vPortSetupTimerInterrupt, which
is only called after a vTaskSchedulerStart() call, allowing it to be
garbage collected if the developer is not using FreeRTOS.