Removing use of env.mk and PATH env variable
1. Replacing env.mk with ~/.sjsu_dev2.mk

Rational:

After taking a deep look at the contents of env.mk I've concluded that
all but the directory location is necessary. The path to the exact
arm-gcc variant can be deduced using the * wildcard character. Same goes
for clang.

Local versions of env.mk has been replaced with a ~/.sjsu_dev2.mk file
located at the HOME directory.

2. Moving away from using soft links

Rational

Soft links break when the source folder moves. To prevent this
a project_makefile.mk is used instead copied into the project folder which
only includes an include for the ~/sjsu_dev2.mk file, which contains the
location of the SJSU-Dev2 directory, then uses that variable to include
the main makefile. This way, moving the SJSU-Dev2 directory does not
break all of the soft links created. The only thing that needs to be
changed is the path in the sjsu_dev2.mk file.

3. Remvoing the use of PATH variable

Rational:

makefile and setup no longer use the PATH environemnt variable, because
there are cases where a user has a globally
installed version of arm-gcc or clang conflicts with the path in the
SJSU-Dev2 folder and the global variant get selected. If the versions
are not the same this could cause undefined behavior.

Resolves #390