Adding dependency injection for newlib
Resolves #401
Resolves #81

main_test.cpp now setups Stdout and Stdin for all tests, removing the
need for macros in the newlib.cpp file to destinguise between host tests
and firmware.

Added injection of Stdout and Stdin in both lpc40xx and lpc17xx.

Added a function to turn off and on _read() echo back to remove the uart
dependancy for commandline.