Add SystemControl as a ctor parameter for LPC40xx L1
By doing this, the SystemController can be dependency injected for
testing and for cross support with the SJOne board.

This change causes issues in areas that rely on the config frequency to
be the actual operating frequency of the system, so the use of
config::kSystemclockRate beyond L0, and some rare exceptions have been
removed.

Resolves issue #551