Fix break in STDIN in newlib
Problems occured when _lseek, _close, _isatty were not defined within
newlib.cpp and where thus defined by rdimon, which would cause a
hardfault when debugger is not plugged in and the system attempts to
read from stdin.

Solution to this was to not define the above functions but to define
their reentrant variants. The reentrant variants which were not hard
defined by rdimon, but get called by rdimon's version of those functions.

Now stdin can be used without a debug port being plugged in.

Also cleanup on log_level_print demo.

Resolves #1037