Esp8266 consistently working
Methods using timeouts now expire properly.

Introducing the TimeoutTimer which takes a timeout and can return the
amount of time left since the TimeoutTimer's creation based on the
systems Uptime() function.

Read no longer waits for multiple +IPD packets. It will simply return
the results of the first +IPD packet it encounters. For HTTP requests,
"Content-Length" should be checked to see if further packets will need
to be read from the stream.