Adding 3P tl::expected library
This library is used to improve the error handling in SJSU-Dev2. It is
an implementation of the std::expected library that was added into the
C++20 standard, but currently is not supported by GCC9. It should exist
in GCC10 and above so when that happens, this library will be removed
from the code base.