Fix libsjsudev2.a missing file build issue
The problem comes down to the fact that multithreading the build can
result in a race condition. In this case, the race condition happens in
this circumstance:

1. Previously built project and all static libraries were generated
2. `make clean`, which deletes the local build file, but not the stored
   static libraries.
3. At this point, there are only two targets to build, the libsjsudev2.a
   combined archive file and the source/main.cpp, so those are built.

What is the race condition?

Currently, the object file build receipe uses  `mkdir -p` to create the
destination directory before generating the file. The libsjsudev2.a
target DOES NOT. If the threads start with libsjsudev2.a before the
main.cpp target has a chance to rebuild the build/objects directory,
the `ar` command invocation will not see an existing directory to
store the file and report that no file or directory exists.

To cause this to happen, do a full build, make clean, then edit the
makefile to remove the main.cpp as a source file to build, such that the
only target left to build is the libsjsudev2.a and this will ALWAYS
fail repeatable.

Resolves #1300