Move CI scripts to CI project
No need to have them take up space in the tools folder. If we can limit
the amount of stuff in that folder and categorically move CI related
stuff to the CI project, this will make the codebase cleaner.

The idea of a scripts folder may also be a good idea for the starter
project since other projects may want to have their own build or
prebuild script or whatever else they want in their projects.

Resolves #1302