Add 3P span library
Add third party span library which is a non-owning container for
contiguous elements. It is a replacement for the need to do such things
as pass a pointer and a size. Spans hold the pointer and size within
themselves and can be efficiently passed around.

NOTE: Keep this until GCC has been upgraded to version 10.

Resolves #901