Reorganize and refactor documenation
Organized the documenation to include an introduction area that include
most of what a beginner wanting to dig into the platform will need to
begin working with SJSU-Dev2. Plan is to keep this as short as possible
and leave the greater details to other bit of documentation.

Documenation is still incomplete for many of these but this lays out the
scaffolding.

Resolves #1347
Resolves #1227
Resolves #1346