Migrate from Travis-CI to Github Workflows
Travis-CI has changed its payment scheme for open source giving a limit
to the number of minutes that can be used for testing. There is, as far
as I know, no way to pay for more credits without subscribing to a
monthly fee. In order to get a continous integration system back, the
decision to move to Github workflows was decided.