StaticAllocator: make ByteAllocated() accurate
Using a variable to accumulate the number of bytes used in each
allocation can give wrong results if the alignment is not taken into
account.

Replacing the acculation variable with a pointer to the unallocated
memory, calculated by taking the most recent allocated address and
adding the number of bytes to that will give you exactly where the end
of the allocated memory is. This takes advantage of the fact that the
std::pmr::monotonic_memory_resource class performs all of the alignment
and allocation math for us.