CanNetwork: Solve shared CAN periperal problem
Add CanNetwork driver which solves a critical issue with the CAN
peripheral API and drivers that must share a single CANBUS/CAN
peripheral object.

The problem with the Can::Receive() method, is that the software
cannot ask/request a message of a specific ID. This is very
important as a L2_HAL driver will only be concerned with specific
messages based on ID.

Most hardware CAN peripherals utilize a small queue that cannot be
searched through. Effectively meaning that CAN messages but be consumed
in sequence. If a driver pulls a message that is not intended for it,
the message must be stored somewhere or dropped. Dropping means that the
driver that wanted that data will not get it, and storage requires that
some system manage the CAN message data.

Dropping is not acceptable, but storage is, thus the CanNetwork was
created.

The CanNetwork stores received CAN messages into a hash map using the
IDs as keys. This "somewhat" requires that the CAN peripherals to
support interrupts when a message is received. Thus the Can interface
was extended to support hardware filtering, enabling the filter and most
importantly a means to set a Receive Handler. The CanNetwork will inject
its own Handler into the Can Peripheral ISR that will capture can
messages received from the CAN BUS provided each ID has been added to
the CanNetwork via the CaptureMessage() method. If an ISR is not
available, application software can drive the handler itself by running
the ManuallyCallReceiveHandler() method directly.

The RMD-X code as been rewritten to use a CanNetwork instead of Can
directly. All L2_HAL drivers must use the CanNetwork going forward.