Add OPTIMIZE to build path
Adding OPTIMIZE to the build path has the sole benefit of removing the
need for users to run `purge` everytime they want to build their
application with a different optimization level. This results in
a quality of life improvement due to:

1. Reduced compile times (can reuse previously generated libraries that
   fit the platform and the optimization level)
2. Improved expectations regarding debugging (compiling with 'g' but
   previously compiled library files were compiled with 3 making
   debugging an unexpected disaster between code that was highly
   optimized and code that is optimized for debugging).

This change does, consume more space user's disk based on the
permutation of platforms and optimization levels used.

Additional Updates:

- At the start of the targets application, flash, and debug a
  "start-message" prompt will appear to give information about the
  target settings such as which PLATFORM, OPTIMIZE, and JTAG used.
- A bit of cleanup
- Documentation updated to reflect these changes to the makefile